requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
bottle==0.12.25
bottle-websocket==0.2.9      
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2    
colorama==0.4.6
Eel==0.16.0
flake8==7.1.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.21
pyflakes==3.2.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.50
trio==0.26.2
typing_extensions==4.12.2
webdriver-manager==4.0.2
websocket-client==1.8.0
whichcraft==0.6.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
